pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "agent", "agentTool", "agentEvaluation"]

[project]
name = "device-ai-agent"
version = "0.1.0"
description = "Agent for control of devices and testing"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0"
]

[project.scripts]